Move casting_replace_symbolt to files of their own
Increase readability of linking code by moving aside code that is not
specific to just linking.
