+++ src/linking/Makefile
@@ -1,4 +1,5 @@
-SRC = linking.cpp \
+SRC = casting_replace_symbol.cpp \
+      linking.cpp \
       remove_internal_symbols.cpp \
       static_lifetime_init.cpp \
       # Empty last line
